login y logout validados con spinner de carga

+++ src/app/layout/navigation/navigation.component.ts
@@ -2,17 +2,22 @@ import { Component, inject } from '@angular/core';
 import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
 import { Observable } from 'rxjs';
 import { map, shareReplay } from 'rxjs/operators';
-import { Router } from '@angular/router'; // Importamos Router
+import { Router } from '@angular/router';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmLogoutDialogComponent } from '../../shared/modales/confirm-logout-login/confirmLogoutDialogComponent';
 
 @Component({
   selector: 'app-navigation',
+  standalone: false,
   templateUrl: './navigation.component.html',
   styleUrl: './navigation.component.scss',
-  standalone: false
+  
 })
 export class NavigationComponent {
   private breakpointObserver = inject(BreakpointObserver);
-  private router = inject(Router); // Inyectamos el Router
+  private router = inject(Router);
+  private dialog = inject(MatDialog); // ✅ Faltaba esta línea
+
 
   isHandset$: Observable<boolean> = this.breakpointObserver.observe(Breakpoints.Handset)
     .pipe(
@@ -22,6 +27,8 @@ export class NavigationComponent {
 
   isSideMode = false;
   sidenavOpened = true;
+
+  cerrandoSesion = false; // 👉 Variable para mostrar el mensaje
 
   constructor() {
     this.isHandset$.subscribe(isHandset => {
@@ -39,13 +46,20 @@ export class NavigationComponent {
       if (isHandset) {
         drawer.close();
       }
-    }).unsubscribe(); // Cerramos la subscripción inmediatamente
+    }).unsubscribe();
   }
 
-  logout() {
-    // Borra datos de sesión
-    localStorage.clear(); // O sessionStorage.clear();
-    // Redirige al login
-    this.router.navigate(['/auth/login']);
-  }
+logout() {
+  const dialogRef = this.dialog.open(ConfirmLogoutDialogComponent);
+
+  dialogRef.afterClosed().subscribe(result => {
+    if (result === true) {
+      this.cerrandoSesion = true;
+      setTimeout(() => {
+        localStorage.clear();
+        this.router.navigate(['/auth/login']);
+      }, 3000);
+    }
+  });
+}
 }
